pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sdpype"
version = "0.1.0"
description = "Synthetic Data Pipeline"
authors = [
    {name = "Aleksandar Babic", email = "aleksandar.babic@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    # Core ML/Data
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.1.0",
    
    # Pipeline & Config
    "dvc>=3.0.0",
    "hydra-core>=1.3.2",
    
    # CLI
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    
    # Synthetic Data Libraries
    "opacus<1.5.0",
    "torch==2.2.2",
    "torchaudio==2.2.2",
    "torchvision==0.17.2",    
    "sdv>=1.0.0",
    "synthcity>=0.2.12",
]

[project.scripts]
sdpype = "sdpype.main:app"
